requirements.txt
requests==2.28.2
python-dotenv==1.0.0
CloudflareBypasser==0.1.7
DrissionPage==1.6.7
selenium==4.15.0
